fix store scripts

+++ jvm/core/main/scala/com/webank/eggroll/core/resourcemanager/job/container/ProcessContainer.scala
@@ -58,12 +58,16 @@ class ProcessContainer(
   }
 
   def stop(): Boolean = {
-    process.destroy()
+    if (process.isAlive) {
+      process.destroy()
+    }
     process.isAlive
   }
 
   def kill(): Boolean = {
-    process.destroyForcibly()
+    if (process.isAlive) {
+      process.destroyForcibly()
+    }
     process.isAlive
   }
